// assets/Models/Prisha.ts
export enum PrishaType {
  VESET_HACHODESH_BEFORE_SUNRISE = "VESET_HACHODESH_BEFORE_SUNRISE",
  VESET_HAFLAGA_BEFORE_SUNRISE = "VESET_HAFLAGA_BEFORE_SUNRISE",
  ONAA_BENONIT_BEFORE_SUNRISE = "ONAA_BENONIT_BEFORE_SUNRISE",

  VESET_HACHODESH_AFTER_SUNSET = "VESET_HACHODESH_AFTER_SUNSET",
  VESET_HAFLAGA_AFTER_SUNSET = "VESET_HAFLAGA_AFTER_SUNSET",
  ONAA_BENONIT_AFTER_SUNSET = "ONAA_BENONIT_AFTER_SUNSET",

  VESET_HACHODESH_DAY = "VESET_HACHODESH_DAY",
  VESET_HAFLAGA_DAY = "VESET_HAFLAGA_DAY",
  ONAA_BENONIT_DAY = "ONAA_BENONIT_DAY",

  OT_ZARIHA_VESET_HACHODESH_BEFORE_SUNRISE = "OT_ZARIHA_VESET_HACHODESH_BEFORE_SUNRISE",
  OT_ZARIHA_VESET_HAFLAGA_BEFORE_SUNRISE = "OT_ZARIHA_VESET_HAFLAGA_BEFORE_SUNRISE",
  OT_ZARIHA_ONAA_BENONIT_BEFORE_SUNRISE = "OT_ZARIHA_ONAA_BENONIT_BEFORE_SUNRISE",

  OT_ZARIHA_VESET_HACHODESH_AFTER_SUNSET = "OT_ZARIHA_VESET_HACHODESH_AFTER_SUNSET",
  OT_ZARIHA_VESET_HAFLAGA_AFTER_SUNSET = "OT_ZARIHA_VESET_HAFLAGA_AFTER_SUNSET",
  OT_ZARIHA_ONAA_BENONIT_AFTER_SUNSET = "OT_ZARIHA_ONAA_BENONIT_AFTER_SUNSET",

  OT_ZARIHA_VESET_HACHODESH_DAY = "OT_ZARIHA_VESET_HACHODESH_DAY",
  OT_ZARIHA_VESET_HAFLAGA_DAY = "OT_ZARIHA_VESET_HAFLAGA_DAY",
  OT_ZARIHA_ONAA_BENONIT_DAY = "OT_ZARIHA_ONAA_BENONIT_DAY"
}

// Mapping each Prisha type to its corresponding number
export const PrishaTimeMapping: Record<PrishaType, number> = {
  [PrishaType.VESET_HACHODESH_BEFORE_SUNRISE]: 1,
  [PrishaType.VESET_HAFLAGA_BEFORE_SUNRISE]: 1,
  [PrishaType.ONAA_BENONIT_BEFORE_SUNRISE]: 1,

  [PrishaType.VESET_HACHODESH_DAY]: 2,
  [PrishaType.VESET_HAFLAGA_DAY]: 2,
  [PrishaType.ONAA_BENONIT_DAY]: 2,

  [PrishaType.VESET_HACHODESH_AFTER_SUNSET]: 3,
  [PrishaType.VESET_HAFLAGA_AFTER_SUNSET]: 3,
  [PrishaType.ONAA_BENONIT_AFTER_SUNSET]: 3,

  [PrishaType.OT_ZARIHA_VESET_HACHODESH_BEFORE_SUNRISE]: 1,
  [PrishaType.OT_ZARIHA_VESET_HAFLAGA_BEFORE_SUNRISE]: 1,
  [PrishaType.OT_ZARIHA_ONAA_BENONIT_BEFORE_SUNRISE]: 1,

  [PrishaType.OT_ZARIHA_VESET_HACHODESH_DAY]: 2,
  [PrishaType.OT_ZARIHA_VESET_HAFLAGA_DAY]: 2,
  [PrishaType.OT_ZARIHA_ONAA_BENONIT_DAY]: 2,

  [PrishaType.OT_ZARIHA_VESET_HACHODESH_AFTER_SUNSET]: 3,
  [PrishaType.OT_ZARIHA_VESET_HAFLAGA_AFTER_SUNSET]: 3,
  [PrishaType.OT_ZARIHA_ONAA_BENONIT_AFTER_SUNSET]: 3,
};